CoreJava_yk/src/com/crmindz/bcjmay2017/Collections/testcases/ReverseOddAndEvenTest.java
package com.crmindz.bcjmay2017.Collections.testcases;


import static org.junit.Assert.assertEquals;

import java.util.ArrayList;

import org.junit.Test;

import com.crmindz.bcjmay2017.Collections.lists.lab1.ReverseOddAndEven;

/**
 * @author Yeshwanth Konakanchi
 * This is the test class for the ReverseOddAndEven class
 */
public class ReverseOddAndEvenTest {

	@Test
	public void testReverseOddAndEvenList() {
		
		ArrayList<String> input = new ArrayList<String>();
		
		input.add("Hello");
		input.add("how");
		input.add("are");
		input.add("you?");
		input.add("My");
		input.add("name");
		input.add("is");
		input.add("Siva");
		input.add("Yarram");
		
		ArrayList<String> expected = new ArrayList<String>();
		
		expected.add("how");
		expected.add("Hello");
		expected.add("you?");
		expected.add("are");
		expected.add("name");
		expected.add("My");
		expected.add("Siva");
		expected.add("is");
		expected.add("Yarram");

		assertEquals(expected,ReverseOddAndEven.reverseOddAndEvenList(input));

	}

}
